requirements.txt
asttokens==2.4.1
blosc2==2.6.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
dask==2024.5.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
fonttools==4.53.0
fsspec==2024.5.0
h5py==3.11.0
hyperspy==2.1.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.4
llvmlite==0.42.0
locket==1.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.0.8
natsort==8.4.0
ndindex==1.8
nest-asyncio==1.6.0
networkx==3.3
numba==0.59.1
numexpr==2.10.0
numpy==1.26.4
packaging==24.0
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==10.3.0
Pint==0.23
pip==24.0
platformdirs==4.2.2
pooch==1.8.1
prettytable==3.10.0
prompt_toolkit==3.0.45
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
Pygments==2.18.0
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
python-box==6.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
requests==2.32.3
rosettasciio==0.4
scikit-image==0.23.2
scipy==1.13.1
setuptools==69.5.1
six==1.16.0
stack-data==0.6.3
sympy==1.12.1
tables==3.9.2
tifffile==2024.5.22
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
traits==6.4.3
typing_extensions==4.12.1
urllib3==2.2.1
wcwidth==0.2.13
wheel==0.43.0
xraylib==4.1.3
zipp==3.19.1
